Function to create HTML EDA report

Create a exploratory data analysis report in HTML format

ExpReport(
  data,
  Template = NULL,
  Target = NULL,
  label = NULL,
  theme = "Default",
  op_file = NULL,
  op_dir = getwd(),
  sc = NULL,
  sn = NULL,
  Rc = NULL
)

Arguments

data
a data frame

Template
R markdown template (.rmd file)

Target
dependent variable. If there is no defined target variable then keep as it is NULL.

label
target variable descriptions, not a mandatory field

theme
customized ggplot theme (default SmartEDA theme) (for Some extra themes use Package: ggthemes)

op_file
output file name (.html)

op_dir
output path

sc
sample number of plots for categorical variable. User can decide how many number of plots to depict in html report.

sn
sample number of plots for numerical variable. User can decide how many number of plots to depict in html report.

Rc
reference category of target variable. If Target is categorical then Pclass value is mandatory and which should not be NULL

Details

The "ExpReport" function will generate a HTML report for any R data frames.

Note

If the markdown template is ready, you can use that template to generate the HTML report

ExpReport will generate three different types of HTML report based on the Target field

IF Target = NULL, means there is no defined dependent varaible then it will genreate general EDA report at overall level

IF Target = continuous, then it will genreate EDA report including univariate and multivarite summary statistics with correlation.

IF Target = categorical, then it will genreate EDA report including univariate and multivarite summary statistics with chi square, Information values.

See also

Examples

## Creating HTML report
if (FALSE) {
 library (ggthemes)
 # Create report where target variable is categorical
 ExpReport(mtcars,Target="gear",label="car",theme=theme_economist(),op_file="Samp1.html",Rc=3)
 # Create report where target variable is continuous
 ExpReport(mtcars,Target="wt",label="car",theme="Default",op_file="Samp2.html")
 # Create report where no target variable defined
 ExpReport(mtcars,Target=NULL,label="car",theme=theme_foundation(),op_file="Samp3.html")
}